Eliminate code dup around item_attrs.
